+++ tree/apps.py
@@ -1,7 +1,13 @@
 from django.apps import AppConfig
 
 from .fields import PathField
-from .lookups import AncestorOf, SiblingOf, ChildOf, DescendantOf
+from .lookups import (
+    AncestorOf,
+    SiblingOf,
+    ChildOf,
+    DescendantOf,
+    StrictDescendantOf,
+)
 from .transforms import Level
 from .types import Path
 
@@ -17,6 +23,7 @@ class TreeAppConfig(AppConfig):
         PathField.register_lookup(SiblingOf)
         PathField.register_lookup(ChildOf)
         PathField.register_lookup(DescendantOf)
+        PathField.register_lookup(StrictDescendantOf)
 
         PathField.register_lookup(Level)